\name{cite.geoR}
\alias{cite.geoR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Citing Package geoR in Publications}
\description{
  How to cite the package \pkg{geoR} in publications.
}
\usage{
cite.geoR()
}
%- maybe also 'usage' for other objects documented here.
\details{
  Execute function \code{cite.geoR()} for information on how to cite geoR in
     publications.
}
%\value{
%  ~Describe the value returned
%  If it is a LIST, use
%  \item{comp1 }{Description of 'comp1'}
%  \item{comp2 }{Description of 'comp2'}
%  ...
%}
%\references{ ~put references to the literature/web site here ~ }
%\author{ ~~who you are~~ }
%\note{ ~~further notes~~ }


%\seealso{ ~~objects to See Also as \code{\link{~~fun~~}}, ~~~ }
\examples{
cite.geoR()
}
\keyword{misc}% at least one, from doc/KEYWORDS
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
